Dapper, Autofac, OWIN

+++ BlackJack.DataAccessLayer/Interfaces/IGameRepository.cs
@@ -4,7 +4,7 @@ namespace BlackJack.DataAccessLayer.Interfaces
 {
     public interface IGameRepository : IBaseRepository<Game>
     {
-        int CreateAndReturnId(Game item);
-        int ReturnNewGameNumber(int id);
+        long CreateAndReturnId(Game item);
+        int ReturnNewGameNumber(long id);
     }
 }
