+++ BlackJack.DataAccessLayer/Interfaces/IRoundRepository.cs
@@ -4,7 +4,7 @@ namespace BlackJack.DataAccessLayer.Interfaces
 {
     public interface IRoundRepository : IBaseRepository<Round>
     {
-        int CreateAndReturnId(Round item);
-        int ReturnNewRoundNumber(int id);
+        long CreateAndReturnId(Round item);
+        long ReturnNewRoundNumber(long id);
     }
 }
