+++ BlackJack.ViewModels/ResponseModel/ResponseGameListHistoryView.cs
@@ -15,14 +15,14 @@ namespace BlackJack.ViewModels.ResponseModel
 
     public class GameGameListHistoryViewItem
     {
-        public int Id { get; set; }
-        public int Number { get; set; }
+        public long Id { get; set; }
+        public long Number { get; set; }
         public int AmountRounds { get; set; }
     }
 
     public class PlayerGameListHistoryViewItem
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
     }
 }
